requirements.txt
# general modules
Django >4.2,<4.3
python-decouple
pillow
djangorestframework


# third party modules
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt


#Email third party module
django-mail-templated

#refactoring and tests
black
flask8
